굿즈 상품판매 게시판 완성

+++ src/main/java/org/webtoon/repository/GoodsRepository.java
@@ -1,10 +1,12 @@
 package org.webtoon.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.webtoon.entity.Goods;
 
-public interface GoodsRepository  extends JpaRepository<Goods,Integer> {
+public interface GoodsRepository  extends JpaRepository<Goods,Long> {
 
+//    @Query("select g from goods g where g.productId=:productId")
     Goods findByProductId(int productId);
